[Refactor] Extract index request factory (#420)
## Problem

With my asyncio client in progress (not in this diff), I found myself
needing to reuse a lot of the openapi request-building boilerplate.

## Solution

Rather than copy/paste and create a lot of duplication, I want to pull
that logic out into a separate class for building request objects. This
change should not break existing behavior in the Index client.

## Todo

This was pretty much just a mechanical extraction. I'd like to cleanup
and standardize some stuff regarding `_check_type`, an openapi param
that we're going to silly lengths to flip the default behavior on, in a
follow-up diff.

## Type of Change

- [x] None of the above: Refactor

## Test Plan

Should still have tests passing