Adjust config kwargs to Pinecone and PineconeAsyncio (#447)
## Problem

We want to deprecate a few kwargs and give more helpful error messages.

## Solution

- Deprecate `config` and `openapi_config` kwargs; as far as I know these
were only used for tests. They have not appeared in documentation. It
adds quite a bit of complexity trying to merge configuration from all
these sources so going forward we would prefer to just expect
configuration from a single source, which are named kwargs.
- Add `NotImplementedError` messages that clearly explain some features
that are not implemented yet for PineconeAsyncio. This is preferrable to
allowing users to pass kwargs that have no effect, then getting
frustrated when they seem not to work.
- Give more informative error when it looks like the user has passed an
invalid value as the host

## Type of Change

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update