Remove explicit tqdm dependency (#448)
## Problem

We want to minimize the number of required dependencies we have, and
tqdm is non-essential. Moreover, common notebook environments like
Google Colab will already have tqdm loaded even if we do not declare
this an explicit dependency.

## Solution

Instead of having a specific dependency on tqdm, we want to detect and
use it if it is available in the environment. Otherwise just noop with a
stub implementation of our own.

## Type of Change

- [x] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [x] This change requires a documentation update