Fix: PineconeGrpcFuture blocks during construction
When using grpc with async_req, the construction of a
PineconeGrpcFuture would call _sync_state, which would do a blocking
call to grpc_future.exception(...). This means that the async_reqs
were all blocking in practice.

We can fix it by checking if the future is still running and not doing
any blocking calls when it is.