Fix backup integration tests (#483)
## Problem

The cleanup steps in the backup tests were deleting all backups in a
project, which creates a problem when multiple jobs in a test matrix are
running in parallel and backups may be suddenly deleted out from
underneath a running test.

## Solution

- Cleanup should only delete backups created with the current RUN_ID.
This information is stored in index tags and is visible on the backup
resource as well.
- Cleanup duplication in conftest setup. It's easier to make these
changes in one spot than in 6-8 spots.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)