Add `py.typed` to indicate type information is present (#485)
## Problem

The Pinecone SDK does not currently have full coverage on mypy type
annotations, but we are adopting them in an incremental fashion.

## Solution

Add a `py.typed` file to the package. An empty `py.typed` file is used
as a
[marker](https://typing.python.org/en/latest/spec/distributing.html#packaging-typed-libraries)
to let mypy or other type checkers know there is type information in the
package.

## Type of Change

- [x] New feature (non-breaking change which adds functionality)

## Test Plan

Describe specific steps for validating this change.