Fix missing readline error on Windows (#503)
## Problem

Missing `readline` dependency error when using `list_models` function.
Reported in #502

## Solution

This error was the result of the Cursor assistant putting some code into
`pinecone/scripts/repl.py` that should only have gone into
`scripts/repl.py` (and not been part of the package distribution).

This wasn't caught in CI because:
- ~`readline` is included in our dev dependencies, which are installed
in CI for testing purposes.~ `readline` is part of the python standard
library, so even though this reference was inadvertantly referenced, the
presence of these import statements should not cause a problem on dev
machines (mac) or CI (linux). However, `readline` is [not available for
Windows](https://github.com/python/cpython/issues/90028) which is why
this error was occurring for the person who reported.
- Currently we do not have any Windows-specific testing in CI which is
why this error shipped unnoticed. So I implemented some new tests to
check the package can be installed on Windows in this PR.

## Type of Change

- [x] Bug fix (non-breaking change which fixes an issue)

## Test Plan

- Going to make a dev build and verify `list_models()` works in google
colab (where dev dependencies are not present) (7.0.2.dev1)
- Tested in Colab with 7.0.0 and 7.0.1 and could not reproduce the
problem; after further research it turns out this is a windows specific
problem. I should probably add some windows builds into the test matrix.
- Added new installation tests workflow to verify Windows install