Only run install and dependency tests on merge (#505)
## Problem

Making changes in this repository sometimes feels slow because of the
amount of tests triggered on every push. Many tests, especially checking
the dependency matrix, are unlikely to fail unless dependencies have
changed so we can run them less often when certain specific conditions
are met.

## Solution

- Instead of testing install and the dependency matrix on every push on
a PR in progress, refactor github workflows to only run these
(`testing-install.yaml`, and `testing-dependency.yaml`) when pushing to
a branch where work is being integrated (e.g. `main` or
`release-candidate/*`).
- Add an additional PR workflow that only triggers when `pyproject.toml`
or `poetry.lock` are modified. When dependencies are changed, we do need
to run the entire dependency testing matrix.

## Type of Change

- [x] Infrastructure change (CI configs, etc)