Fix release workflow (#516)
## Problem

When I tried releasing to prod, the workflow failed with the following
error:
```
[Invalid workflow file: .github/workflows/release-prod.yaml#L26](https://github.com/pinecone-io/pinecone-python-client/actions/runs/15689847725/workflow)
The workflow is not valid. .github/workflows/release-prod.yaml (Line: 26, Col: 11): Input python_versions_json is required, but not provided while calling.
```

## Solution

Recently, the unit and integration tests workflows were updated in this
pr for creating a new project for each test run. These workflows are
used in release-prod.yaml but the corresponding changes were not made.
As a part of this PR, I have:
1. added `python_versions_json` for running unit-tests (required by
[test-unit.yaml](https://github.com/pinecone-io/pinecone-python-client/pull/506/files#diff-470b33906ddc2d9d1d4ef14595bfb97dabdc66146282122f5f8109e77a00f1eaR5))
2. added `encrypted_project_api_key` and `python_versions_json` for
running integration tests (required by
[test-integration.yaml](https://github.com/pinecone-io/pinecone-python-client/pull/506/files#diff-3aae3c6c7bf80d1ee8f43fe2844445abb98c3792ed1da99ef5fff84ce330e549R5))
3. Added create-project and cleanup-project as extra steps (based on
[this
pr](https://github.com/pinecone-io/pinecone-python-client/pull/506))


## Type of Change

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] This change requires a documentation update
- [X] Infrastructure change (CI configs, etc)
- [ ] Non-code change (docs, etc)
- [ ] None of the above: (explain here)

## Test Plan

NA