Speed up grpc fetch and query response parsing
Running a profiler on my pinecone program, it was CPU-bottlenecked
on json_format.MessageToDict (being able to do about 100 vectors per
second in query and fetch responses).
It turns out converting the embeddings to a dict this way is very
slow. It's much faster to convert them to a list without going through
MessageToDict.